+++ test/resources/checkers/sh-shellcheck.sh
@@ -1,3 +1,4 @@
 #!/bin/sh
 rm "~/my file.txt"
 touch $@
+export LS=`which ls`